+++ lororUtil/src/main/java/com/loror/lororUtil/http/ProgressListener.java
@@ -1,7 +1,7 @@
 package com.loror.lororUtil.http;
 
 public interface ProgressListener {
-	void transing(int progress, int speed, long length);
+	void transing(float progress, int speed, long length);
 
 	void failed();
